Misc code improvements

+++ exist-core/src/main/java/org/exist/xmldb/AbstractRemoteResource.java
@@ -262,8 +262,8 @@ public abstract class AbstractRemoteResource extends AbstractRemote
         final List<Object> params = new ArrayList<>();
         if (isRetrieve) {
             command = "retrieveFirstChunk";
-            params.add(Integer.valueOf(handle));
-            params.add(Integer.valueOf(pos));
+            params.add(handle);
+            params.add(pos);
         } else {
             command = "getDocumentData";
             params.add(path.toString());
@@ -287,7 +287,7 @@ public abstract class AbstractRemoteResource extends AbstractRemote
                 method = "getNextChunk";
             }
 
-            long offset = ((Integer) table.get("offset")).intValue();
+            long offset = (Integer) table.get("offset");
             byte[] data = (byte[]) table.get("data");
             final boolean isCompressed = "yes".equals(properties.getProperty(EXistOutputKeys.COMPRESS_OUTPUT, "no"));
 
@@ -361,7 +361,7 @@ public abstract class AbstractRemoteResource extends AbstractRemote
         if (inMemoryBufferSize == null) {
             inMemoryBufferSize = new LazyVal<>(() -> Integer.parseInt(properties.getProperty("in-memory-buffer-size", Integer.toString(VirtualTempPath.DEFAULT_IN_MEMORY_SIZE))));
         }
-        return inMemoryBufferSize.get().intValue();
+        return inMemoryBufferSize.get();
     }
 
     protected static InputStream getAnyStream(final Object obj)
